Refactoring to support better sequential circuit translation
